revert moving fixtures to separate file
