+++ .flake8
@@ -3,5 +3,3 @@ ignore = E,F403,F405,F541,F841,W
 select = E9,F,W6
 per-file-ignores =
     __init__.py:F401
-    tests/unit/inference/test_inference.py:F811
-    tests/unit/inference/test_model_profiling.py:F811
